Crate sel4_panicking_env

Macros

abort

Aborts execution with a message.

debug_print

Like std::print!, except backed by debug_put_char.

debug_println

Like std::println!, except backed by debug_put_char.

register_abort_hook

Registers an abort hook to be used by abort! and abort_without_info.

register_debug_put_char

Registers a function to be used by debug_put_char, debug_print!, and debug_println!.

Structs

AbortInfo

Information about an abort passed to an abort hook.

Functions

abort_without_info

Aborts without any AbortInfo.

debug_put_char

Prints via a link-time hook.